Get the ranking list

get_event returns a dataframe containing the ranking list for a given event, agegroup, gender and year. Optionally, top_n can be specified to only return the first n performances

get_event(event, agegroup = "ALL", gender = "M", year = "2020", top_n = NULL)

Arguments

event
character string. The event, can be one of: "60", "100", "200", "400", "800", "1500", "3000", "5000", "10000","3000SC", "10K", "HM", "Mar", "60H", "110H", "400H", "HJ", "PV","LJ", "TJ", "SP7.26K", "DT2K", "HT7.26K", "JT800", "HepI", "Dec", "20KW", "50KW", "4x100" or "4x400".

agegroup
character string. Specify age group e.g., "U20", defaults to "ALL".

gender
character. Either "M" for men or "W" for women.

year
character. The year to return in YYYY format, e.g., "2020".

top_n
integer. Return top n rows.

Examples

if (FALSE) {

get_event(event = "100", agegroup = "ALL", gender = "M", year = "2016", top_n = 10)
}